Refactoring and cleaning code

+++ include/renderer.hpp
@@ -2,6 +2,7 @@
 
 #if NEURO_CAR_GRAPHIC_MODE_SFML
 
+#include <cstdint>
 #include <vector>
 
 #include <SFML/Graphics.hpp>
@@ -11,16 +12,16 @@ class Drawable;
 class Renderer
 {
 public:
-    Renderer(int scale, int width, int height);
+    Renderer(uint32_t scale, uint32_t width, uint32_t height);
     ~Renderer();
 
     bool update(std::vector<Drawable*> const & actorList, bool draw = true);
 
 protected:
     sf::RenderWindow m_window;
-    float m_scale; // pixels per meter
-    int m_width;
-    int m_height;
+    uint32_t m_scale; // pixels per meter
+    uint32_t m_width;
+    uint32_t m_height;
 };
 
 #endif // NEURO_CAR_GRAPHIC_MODE_SFML
